fix: extraction
